+++ NEWS.md
@@ -8,6 +8,7 @@ difference formulation thanks to the case study of Mitzi Morris.
 ### Other changes 
 
 * Use hashes to check if models have the same response values when performing model comparisons. (#414)
+* No longer set `pointwise` dynamically in `loo` and related methods. (#416)
 
 * No longer show information criteria in the summary output.